[#3] teacher student reformat
