+++ mnist/utils/models.py
@@ -1,0 +1,34 @@
+import torch.nn as nn
+import torch.nn.functional as F
+
+
+class TeacherNet(nn.Module):
+    def __init__(self):
+        super(TeacherNet, self).__init__()
+        self.fc1 = nn.Linear(28 * 28, 1200)
+        self.fc2 = nn.Linear(1200, 1200)
+        self.fc3 = nn.Linear(1200, 10)
+
+    def forward(self, x):
+        x = x.reshape(-1, 28 * 28)
+        x = F.relu(self.fc1(x))
+        x = F.dropout(x, p=0.8, training=self.training)
+        x = F.relu(self.fc2(x))
+        x = F.dropout(x, p=0.8, training=self.training)
+        x = self.fc3(x)
+        return x
+
+
+class StudentNet(nn.Module):
+    def __init__(self):
+        super(StudentNet, self).__init__()
+        self.fc1 = nn.Linear(28 * 28, 800)
+        self.fc2 = nn.Linear(800, 800)
+        self.fc3 = nn.Linear(800, 10)
+
+    def forward(self, x):
+        x = x.reshape(-1, 28 * 28)  # flatten (N, C, H, W) --> (N, L)
+        x = F.relu(self.fc1(x))
+        x = F.relu(self.fc2(x))
+        x = self.fc3(x)
+        return x
